requirements.txt
pyautogui==0.9.53
Pillow==8.4.0
pytesseract==0.3.8